// finance_frontend/src/api/models/TransactionTypeEnum.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * * `ETC_INCOME` - 기타소득
 * * `SALARY` - 월급
 * * `BONUS` - 보너스
 * * `INTEREST_INCOME` - 이자소득
 * * `STOCK_INCOME` - 주식
 * * `RENT_INCOME` - 임대
 * * `DIVIDEND` - 배당
 * * `DEPOSIT` - 입금
 * * `ETC_EXPENSE` - 기타지출
 * * `DAILY_NECESSITY` - 생필품
 * * `GROCERY` - 식료품
 * * `MEMBERSHIP` - 맴버십
 * * `SERVICE` - 서비스
 * * `EAT_OUT` - 외식
 * * `CLOTHING` - 옷
 * * `PRESENT` - 선물
 * * `CAR` - 이동수단
 * * `HOUSING` - 월세
 * * `LEISURE` - 여가
 * * `MEDICAL` - 의료
 * * `PARENTING` - 육아
 * * `INTEREST_EXPENSE` - 이자지출
 * * `WITHDRAW` - 출금
 * * `TRANSFER` - 이체
 * * `UNKNOWN` - 알 수 없음
 */
export enum TransactionTypeEnum {
  ETC_INCOME = 'ETC_INCOME',
  SALARY = 'SALARY',
  BONUS = 'BONUS',
  INTEREST_INCOME = 'INTEREST_INCOME',
  STOCK_INCOME = 'STOCK_INCOME',
  RENT_INCOME = 'RENT_INCOME',
  DIVIDEND = 'DIVIDEND',
  DEPOSIT = 'DEPOSIT',
  ETC_EXPENSE = 'ETC_EXPENSE',
  DAILY_NECESSITY = 'DAILY_NECESSITY',
  GROCERY = 'GROCERY',
  MEMBERSHIP = 'MEMBERSHIP',
  SERVICE = 'SERVICE',
  EAT_OUT = 'EAT_OUT',
  CLOTHING = 'CLOTHING',
  PRESENT = 'PRESENT',
  CAR = 'CAR',
  HOUSING = 'HOUSING',
  LEISURE = 'LEISURE',
  MEDICAL = 'MEDICAL',
  PARENTING = 'PARENTING',
  INTEREST_EXPENSE = 'INTEREST_EXPENSE',
  WITHDRAW = 'WITHDRAW',
  TRANSFER = 'TRANSFER',
  UNKNOWN = 'UNKNOWN',
}